feat: 로또 객체 생성(generateLottos() 로또객체 내부로 이동)
