+++ README.md
@@ -16,7 +16,7 @@ domain/
 
 Lotto.js
 
-- 중복되지않은 6개의 1~45 랜덤 숫자를 발행
+- ✅ 중복되지않은 6개의 1~45 랜덤 숫자를 발행
 
 Winnings.js
